% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_period.R
\name{edu_period}
\alias{edu_period}
\title{Educational period}
\usage{
edu_period(x)
}
\arguments{
\item{x}{Date or POSIXct vector or value}
}
\value{
the translated period as an integer vector or value
}
\description{
Translate date to educational period
}
\examples{

}
