% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_period.R
\name{edu_period_year}
\alias{edu_period_year}
\title{Educational period}
\usage{
edu_period_year(x, type, name)
}
\description{
This function translates the date vector to a vector of academic periods or
years
}
