% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_period.R
\name{edu_year}
\alias{edu_year}
\title{Educational year}
\usage{
edu_year(x)
}
\arguments{
\item{x}{Date or POSIXct vector or value}
}
\value{
the translated year as an integer vector or value
}
\description{
Translate date to educational year
}
\examples{

}
