% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_dates_csv}
\alias{process_dates_csv}
\title{Process dates csv}
\usage{
process_dates_csv(file, delim = ",", format = "\%d-\%m-\%Y",
  register = TRUE, ...)
}
\arguments{
\item{file}{the path of the csv-file}

\item{delim}{the delimiter default: ","}

\item{format}{the format of the date field default: "\%d-\%m-\%Y"}

\item{register}{default: TRUE register the dates csv in the package options}

\item{...}{extra parameters passed to readr read_delim}
}
\description{
Process a csv file to the right format. This file is used to know the start
and ending dates of the educational periods. The format of the file is.
}
