% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_dates.R
\name{register_dates}
\alias{register_dates}
\title{Register the dates file}
\usage{
register_dates(name = NULL, dates = NULL)
}
\arguments{
\item{name}{use one of the default dates files for example "VU"}

\item{dates}{use a dates table}
}
\description{
Register the dates file
}
