% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edu_period.R
\name{edu_year}
\alias{edu_year}
\title{Educational year}
\usage{
edu_year(x, dates = NULL)
}
\arguments{
\item{x}{Date or POSIXct vector or value}

\item{dates}{optional: use a custom dates table}
}
\value{
the translated year as an integer vector or value
}
\description{
Translate date to educational year
}
\examples{

}
